+++ pycommon_database/flask_restplus_errors.py
@@ -10,6 +10,11 @@ class ValidationFailed(Exception):
         self.errors = marshmallow_errors
 
 
+class ModelCouldNotBeFound(Exception):
+    def __init__(self, requested_data):
+        self.requested_data = requested_data
+
+
 def _failed_field_validation_model(api):
     exception_details = {
         'field_name': fields.String(description='Name of the field that could not be validated.',
@@ -52,3 +57,28 @@ def add_failed_validation_handler(api):
         return {'fields': error_list}, 400
 
     return 400, 'Validation failed.', exception_model
+
+
+def _model_could_not_be_found_model(api):
+    exception_details = fields.String(description='Description of the error.',
+                                      required=True,
+                                      example='Corresponding model could not be found.')
+    return api.model('ModelCouldNotBeFound', exception_details)
+
+
+def add_model_could_not_be_found_handler(api):
+    """
+    Add the default ModelCouldNotBeFound handler.
+
+    :param api: The root Api
+    """
+    exception_model = _model_could_not_be_found_model(api)
+
+    @api.errorhandler(ModelCouldNotBeFound)
+    @api.marshal_with(exception_model, code=404)
+    def handle_exception(model_could_not_be_found):
+        """This is the default model could not be found handling."""
+        logger.exception('Corresponding model could not be found.')
+        return 'Corresponding model could not be found.', 404
+
+    return 404, 'Corresponding model could not be found.', exception_model